refactor(net): 适配跨平台，统一套接字抽象与Windows IOCP支持
1.  统一使用intptr_t替代int作为套接字类型，适配Windows SOCKET类型
2.  新增Platform.h抽象层，封装Windows/Linux平台差异
3.  实现Windows IOCP Poller，完成Windows平台网络IO支持
4.  重构SocketOps、EventLoop等模块适配新平台抽象
5.  清理.gitignore冗余配置，新增测试文件与Windows编译支持

+++ muduox/net/Acceptor.cpp
@@ -34,7 +34,7 @@ void Acceptor::listen() {
 
 void Acceptor::handleRead() {
     InetAddress peerAddr;
-    int connfd = acceptSocket_.accept(&peerAddr);
+    intptr_t connfd = acceptSocket_.accept(&peerAddr);
     if (connfd >= 0) {
         if (newConnectionCallback_) {
             newConnectionCallback_(connfd, peerAddr);
